refactor(provider): rename minUpdateIntervalSec to minUpdateIntervalSeconds
Why: Improved clarity and consistency by renaming `minUpdateIntervalSec` to `minUpdateIntervalSeconds` across `ProviderSettings` and related classes. Ensures alignment with naming conventions and reduces ambiguity in interval representation.

+++ backend/src/main/java/com/alligator/market/backend/provider/catalog/settings/immutable/persistence/jpa/ProviderSettingsEntity.java
@@ -40,9 +40,9 @@ public class ProviderSettingsEntity extends BaseEntity {
     @Column(name = "provider_code", length = 50, nullable = false, updatable = false)
     private String providerCode;
 
-    /** Минимальный интервал обновления котировки в секундах {@link ProviderSettings#minUpdateIntervalSec()}. */
+    /** Минимальный интервал обновления котировки в секундах {@link ProviderSettings#minUpdateIntervalSeconds()}. */
     @NotNull
     @Min(1)
     @Column(name = "min_update_interval_seconds", nullable = false, updatable = false)
-    private Long minUpdateIntervalSeconds;
+    private Integer minUpdateIntervalSeconds;
 }
